html

+++ week-07/day-2/src/main/java/com/greenfoxacademy/simbank/Model/BankAccount.java
@@ -3,10 +3,10 @@ package com.greenfoxacademy.simbank.Model;
 public class BankAccount {
 
     private String name;
-    private int balance;
+    private double balance;
     private String animalType;
 
-    public BankAccount(String name, int balance, String animalType) {
+    public BankAccount(String name, double balance, String animalType) {
         this.name = name;
         this.balance = balance;
         this.animalType = animalType;
@@ -15,8 +15,8 @@ public class BankAccount {
     public String getName() {
         return name;
     }
-    
-    public int getBalance() {
+
+    public double getBalance() {
         return balance;
     }
